[#67]fix: build에러 | suspnse 추가
